sink_connector/requirements.txt
annotated-types==0.7.0
anyio==4.6.0
arrow==1.3.0
asttokens==2.4.1
attrs==24.2.0
autocommand==2.2.2
autotools==0.7
avro==1.12.0
backcall==0.2.0
backports.tarfile==1.2.0
beautifulsoup4==4.12.3
bison==0.1.3
bleach==6.1.0
bz2file==0.98
cdk==1.0.12
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
configparser==7.1.0
confluent-kafka==2.6.0
contourpy==1.3.0
cryptography==43.0.1
cycler==0.12.1
decorator==5.1.1
defusedxml==0.7.1
distro==1.9.0
dnspython==2.7.0
docopt==0.6.2
executing==2.1.0
fastavro==1.9.7
fasteners==0.19
fastjsonschema==2.20.0
filelock==3.16.1
flatten-json==0.1.14
flex==6.14.1
fonttools==4.54.1
fsspec==2024.9.0
gcloud==0.18.3
gevent==24.10.2
googleapis-common-protos==1.65.0
greenlet==3.1.1
grpcio==1.66.2
gTTS==2.5.3
h11==0.14.0
httpcore==1.0.6
httplib2==0.22.0
httpx==0.27.2
idna==3.10
importlib_metadata==8.5.0
ipython==8.12.3
irc==20.5.0
jaraco.classes==3.4.0
jaraco.collections==5.1.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
jaraco.logging==3.3.0
jaraco.stream==3.0.4
jaraco.text==4.0.0
jedi==0.19.1
jeepney==0.8.0
Jinja2==3.1.4
jinja2-time==0.2.0
jiter==0.6.1
jmespath==1.0.1
jolt==0.9.301
json-normalize==1.1.0
jsoncut==0.6
jsonflatten==0.2
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
jwcrypto==1.5.6
kafka==1.3.5
kafka-python==2.0.2
keyring==25.4.1
keyrings.alt==5.0.2
kiwisolver==1.4.7
konfig==1.1
loads==0.3
lxml==5.3.0
make==0.1.6.post2
markdown-it-py==3.0.0
MarkupSafe==3.0.1
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.0.2
more-itertools==10.5.0
mpmath==1.3.0
msys2dl==2.1.1
multi_key_dict==2.0.3
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
networkx==3.4.1
ninja==1.11.1.1
numpy==2.1.2
oauth2client==4.1.3
openai==1.51.2
opencv-python==4.10.0.84
packaging==24.1
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
pendulum==3.0.0
pickleshare==0.7.5
pillow==10.4.0
pip-check-reqs==2.5.3
pipreqs==0.5.0
platformdirs==4.3.6
playsound==1.3.0
prompt_toolkit==3.0.48
protobuf==5.28.2
psutil==6.0.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pycryptodome==3.21.0
pydantic==2.9.2
pydantic_core==2.24.1
pygit2==1.16.0
Pygments==2.18.0
pyjnius==1.6.1
pymongo==4.10.1
pyparsing==3.1.4
Pyrebase4==4.8.0
python-dateutil==2.9.0.post0
python-jwt==4.1.0
pytz==2024.2
pywin32==307
pywin32-ctypes==0.2.3
PyYAML==6.0.2
pyzmq==26.2.0
referencing==0.35.1
requests==2.32.3
requests-toolbelt==1.0.0
rfc3987==1.3.8
rich==13.9.2
rpds-py==0.20.0
rsa==4.9
scipy==1.14.1
seaborn==0.13.2
SecretStorage==3.3.3
setuptools==75.1.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
SpeechRecognition==3.10.4
SQLAlchemy==2.0.35
stack-data==0.6.3
strict-rfc3339==0.7
sympy==1.13.3
tempora==5.7.0
time-machine==2.16.0
tinycss2==1.3.0
torch==2.4.1
torchvision==0.19.1
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
types-python-dateutil==2.9.0.20241003
typing_extensions==4.12.2
tzdata==2024.2
ujson==5.10.0
ultralytics==8.3.10
ultralytics-thop==2.0.9
urllib3==2.2.3
validate_email==1.3
waitress==3.0.0
wcwidth==0.2.13
webencodings==0.5.1
WebOb==1.8.8
WebTest==3.0.1
ws4py==0.5.1
WSGIProxy2==0.5.1
yarg==0.1.9
zipp==3.20.2
zope.event==5.0
zope.interface==7.1.0
zstandard==0.23.0
